patterns completed
