--- DSA - Practice/practice codes/patterns/pattern1/playground.txt
@@ -1,5 +1,0 @@
-1 1 1
-2 2 2
-3 3 3
-
-in every row, we are displaying the `row+1`
