+++ DSA - Practice/practice codes/patterns/pattern2/playground.txt
@@ -1,0 +1,1 @@
+
