--- DSA - Practice/practice codes/patterns/pattern5/code.cpp
@@ -1,8 +1,0 @@
-#include<bits/stdc++.h>
-using namespace std;
-
-int main () {
-
-
-    return 0;
-}
