--- DSA - Practice/practice codes/patterns/pattern5/notes.md
@@ -1,1 +1,0 @@
-#Notes
